Algorithms from past lecture
Found in a script helpfull codes for our project:
One is the effective-distance function called 'Floyd-Warshall'-algorithm
(doesn't do yet exactly what we need but does the hard work of it) and
then I also implemented in a way the 'depth-first search' to check if a
network is still connected.
With thoes two function we can really investigate the different
networks!