Results
Added 2 new folders with results:
* Folder: Characterstics
Implemented all the parameter functions and calculated the values and
put them in a table in the latex file
* Folder: Camparison with real values
The title says it all. But the values as bmatrices in the latex file.

Both thing have still be to be commented on.
But now I'm too tired^^